fixed major bug in quality checks + slightly safer topic term recording
Code was broken if a quality check wasn't in place (d'oh). Fixed this
and decided to record a QualityCheck with a null term so that multiple
refreshes couldn't create a QualityCheck for each document.

Also added a check that topic terms were included in the list of
document terms provided to users to prevent selecting topic terms that
weren't options.