major fix to work with multiple users
This is really bad. Prior versions treated the Document model as though
it was assigned to only one user. So any shared documents (e.g. IRR
docs) were marked as complete for all users after only one user
annotated.

This fix adds a DocumentAssignment model to separate out skippability
and completedness of documents on a per-user basis. This change
necessitated several other changes to other files, which should be self
explanatory.

Additionally:
- Now doesn't delete prior annotations until just before adding new ones
to reduce the chance of completely erasing and failing to update.
- Better wording of "ask for more docs" text.
- Fixed disabling of additional terms when a document is maxed out and
revisited after completion.