quick and dirty sandboxing of documents
Documents are now loaded in an iframe pointing to a separate document
page on the server (/document/<docID>). This ensures that the HTML
describing the document is isolated (in the iframe) from the HTML for
the annotation page.

Because of the random injection of quality checks, this quick solution
involves repeated code when loading either the annotation page or the
raw document page. This ensures that the quality check is injected even
if the document is somehow initially loaded on the document page instead
of the annotation page, but a better solution must exist.